infrastructure/pyproject.toml
[project]
name = "bda-infrastructure"
version = "0.1.0"
dependencies = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
    "aws-cdk.aws-lambda-python-alpha>=2.100.0a0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stacks"]
